+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 7.5.0
 
-- Fixed bugs in the implementation of the `TaskCompletionSource` class.
+- Fixed a minor bug in the implementation of the `TaskCompletionSource` class.
+- Fixed a minor bug in the implementation of the `Task.whenAll()` method.
+- Fixed a minor bug in the implementation of the `Task.whenEach()` method.
+- Fixed a minor bug in the implementation of the `ZonedWork._createPeriodicTimer()` method.
+- Fixed a bug in the implementation of the `ZonedWork.current` getter.
 - Added test file `task_completion_source_test.dart`.
 
 ## 7.4.0